backend/requirement.txt
flask
flask_restx
flask_sqlalchemy
flask_jwt_extended
python_decouple
psycopg2-binary
python-dotenv